# Speed up DogStatsd._report()