@yannmh 2017-10-30
Thanks a lot @shargan for the detailed content on your PR 🙇 

The changes look great and I am looking forward to see them included in the next release!